feat: add mypy type checking and fix test failures
Type Checking:
- Add mypy v1.18.2 as dev dependency
- Integrate mypy into CI workflow (optimize installation step)
- Add mypy pre-commit hook (src/ only) for automatic type checking
- Add 'just typecheck' command and integrate into 'just prepare'
- Fix NumPy type errors by casting to float():
  - src/domain/tuning.py:87 (frequency calculation)
  - src/acoustics/roughness.py:116 (dissonance curve)

Test Fixes:
- Fix regex metacharacter escaping in test_harmonics.py
- Escape square brackets in error message patterns
- All 126 tests now pass, coverage remains at 100%

+++ src/domain/tuning.py
@@ -84,7 +84,7 @@ class TuningSystem:
             >>> tuning.get_frequency(-12) # A3 (octave down)
             220.0
         """
-        return self.base_frequency * (2.0 ** (step / self.edo))
+        return float(self.base_frequency * (2.0 ** (step / self.edo)))
 
     def get_interval_cents(self, steps: int) -> float:
         """指定されたステップ数のインターバルサイズをセントで計算します。
